Add StringBuilderExtensions.cs with StringBuilder extension methods
